Build raylib static library as part of the build for the executable
We cannot use the release build of libraylib.a because it is built with
target `arm64-apple-macos11` which adds msgSend optimizations that are
not compatible with the linker (clang 15) that's used by leanc.

We must add the compiler flag `-fno-objc-msgsend-selector-stubs` to make the
library compatible with the leanc linker.

See https://stackoverflow.com/a/74054943