Use `LEAN_CC=clang lake` for builds on linux.
The LEAN_CC variable was introduced recently to use override the bundled
clang (leanc) from the lean toolchain.

https://github.com/leanprover/lean4/pull/4477#issuecomment-2174623805

We must use LEAN_CC on linux because the example executable will fail to
link if raylib is built with a glibc that's incompatible with the version
that's bundled with leanc. This currently happens with the arch distro.

Thanks to @keeganperry7 for figuring this out.

Co-authored-by: keeganperry7 <keegan.perry.kp@gmail.com>